Seperated server configuration from request handling
