+++ src/tests/grpc/send_query.rs
@@ -1,10 +1,10 @@
 #[cfg(test)]
 mod refinements {
-    use crate::ProtobufServer::server::ConcreteEcdarBackend;
     use crate::ProtobufServer::services::component::Rep;
     use crate::ProtobufServer::services::ecdar_backend_server::EcdarBackend;
     use crate::ProtobufServer::services::query_response;
     use crate::ProtobufServer::services::{Component, ComponentsUpdateRequest, Query};
+    use crate::ProtobufServer::ConcreteEcdarBackend;
     use tonic::Request;
 
     static CONJUN: &str = "samples/xml/conjun.xml";
